# web/project-site/public/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Audio transcription service using WhisperX",
        "title": "Scriberr API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/summarize": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Stream an LLM-generated summary for provided content; persists latest summary for the transcription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "summarize"
                ],
                "summary": "Summarize content",
                "parameters": [
                    {
                        "description": "Summarize request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.SummarizeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event stream",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/transcription/{id}/summary": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the most recent saved summary for the given transcription",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "summarize"
                ],
                "summary": "Get latest summary for transcription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transcription ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Summary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.SummarizeRequest": {
            "type": "object",
            "required": [
                "content",
                "model",
                "transcription_id"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "transcription_id": {
                    "type": "string"
                }
            }
        },
...
        "models.JobStatus": {
            "type": "string",
            "enum": [
                "uploaded",
                "pending",
                "processing",
                "completed",
                "failed"
            ],
            "x-enum-varnames": [
                "StatusUploaded",
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusFailed"
            ]
        },
        "models.MultiTrackFile": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "file_name": {
                    "description": "Original filename (used as speaker name)",
                    "type": "string"
                },
                "file_path": {
                    "description": "Full path to audio file",
                    "type": "string"
                },
                "gain": {
                    "description": "Gain value from .aup file",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "mute": {
                    "description": "Whether track is muted",
                    "type": "boolean"
                },
                "offset": {
                    "description": "Offset in seconds from .aup file",
                    "type": "number"
                },
                "pan": {
                    "description": "Pan value from .aup file (-1.0 to 1.0)",
                    "type": "number"
                },
                "track_index": {
                    "description": "Order of the track",
                    "type": "integer"
                },
                "transcription_job": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TranscriptionJob"
                        }
                    ]
                },
                "transcription_job_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Summary": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "template_id": {
                    "type": "string"
                },
                "transcription": {
                    "description": "Relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.TranscriptionJob"
                        }
                    ]
                },
                "transcription_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.TranscriptionJob": {
            "type": "object",
            "properties": {
                "audio_path": {
                    "type": "string"
                },
                "aup_file_path": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "diarization": {
                    "type": "boolean"
                },
                "error_message": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "individual_transcripts": {
                    "description": "JSON-serialized map[string]*string",
                    "type": "string"
                },
                "is_multi_track": {
                    "type": "boolean"
                },
                "merge_error": {
                    "type": "string"
                },
                "merge_status": {
                    "description": "none, pending, processing, completed, failed",
                    "type": "string"
                },
                "merged_audio_path": {
                    "type": "string"
                },
                "multi_track_files": {
                    "description": "Relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.MultiTrackFile"
                    }
                },
                "multi_track_folder": {
                    "type": "string"
                },
                "parameters": {
                    "description": "WhisperX parameters",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.WhisperXParams"
                        }
                    ]
                },
                "pinned_execution_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.JobStatus"
                },
                "summary": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "transcript": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.WhisperXParams": {
            "type": "object",
            "properties": {
                "align_model": {
                    "description": "Alignment settings",
                    "type": "string"
                },
                "api_key": {
                    "description": "OpenAI settings",
                    "type": "string"
                },
                "attention_context_left": {
                    "description": "NVIDIA model-specific parameters",
                    "type": "integer"
                },
                "attention_context_right": {
                    "type": "integer"
                },
                "batch_size": {
                    "type": "integer"
                },
                "beam_size": {
                    "type": "integer"
                },
                "best_of": {
                    "type": "integer"
                },
                "callback_url": {
                    "description": "Webhook settings",
                    "type": "string"
                },
                "chunk_size": {
                    "type": "integer"
                },
                "compression_ratio_threshold": {
                    "type": "number"
                },
                "compute_type": {
                    "type": "string"
                },
                "condition_on_previous_text": {
                    "type": "boolean"
                },
                "device": {
                    "description": "Device and computation",
                    "type": "string"
                },
                "device_index": {
                    "type": "integer"
                },
                "diarize": {
                    "description": "Diarization settings",
                    "type": "boolean"
                },
                "diarize_model": {
                    "description": "Options: 'pyannote', 'nvidia_sortformer'",
                    "type": "string"
                },
                "fp16": {
                    "type": "boolean"
                },
                "hf_token": {
                    "description": "Token and progress",
                    "type": "string"
                },
                "highlight_words": {
                    "type": "boolean"
                },
                "initial_prompt": {
                    "type": "string"
                },
                "interpolate_method": {
                    "type": "string"
                },
                "is_multi_track_enabled": {
                    "description": "Multi-track transcription settings",
                    "type": "boolean"
                },
                "language": {
                    "type": "string"
                },
                "length_penalty": {
                    "type": "number"
                },
                "logprob_threshold": {
                    "type": "number"
                },
                "max_line_count": {
                    "type": "integer"
                },
                "max_line_width": {
                    "description": "Output formatting",
                    "type": "integer"
                },
                "max_new_tokens": {
                    "description": "Voxtral settings",
                    "type": "integer"
                },
                "max_speakers": {
                    "type": "integer"
                },
                "min_speakers": {
                    "type": "integer"
                },
                "model": {
                    "description": "Model parameters",
                    "type": "string"
                },
                "model_cache_only": {
                    "type": "boolean"
                },
                "model_dir": {
                    "type": "string"
                },
                "model_family": {
                    "description": "Model family (whisper or nvidia)",
                    "type": "string"
                },
                "no_align": {
                    "type": "boolean"
                },
                "no_speech_threshold": {
                    "type": "number"
                },
                "nvidia_chunk_duration": {
                    "type": "integer"
                },
                "nvidia_precision": {
                    "type": "string"
                },
                "nvidia_prompt": {
                    "type": "string"
                },
                "nvidia_target_language": {
                    "type": "string"
                },
                "nvidia_timestamps": {
                    "type": "boolean"
                },
                "nvidia_use_chunking": {
                    "type": "boolean"
                },
                "output_format": {
                    "description": "Output settings",
                    "type": "string"
                },
                "patience": {
                    "type": "number"
                },
                "print_progress": {
                    "type": "boolean"
                },
                "return_char_alignments": {
                    "type": "boolean"
                },
                "segment_resolution": {
                    "type": "string"
                },
                "speaker_embeddings": {
                    "type": "boolean"
                },
                "suppress_numerals": {
                    "type": "boolean"
                },
                "suppress_tokens": {
                    "type": "string"
                },
                "task": {
                    "description": "Task and language",
                    "type": "string"
                },
                "temperature": {
                    "description": "Transcription quality settings",
                    "type": "number"
                },
                "temperature_increment_on_fallback": {
                    "type": "number"
                },
                "threads": {
                    "type": "integer"
                },
                "vad_method": {
                    "description": "VAD (Voice Activity Detection) settings",
                    "type": "string"
                },
                "vad_offset": {
                    "type": "number"
                },
                "vad_onset": {
                    "type": "number"
                },
                "verbose": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token with Bearer prefix",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}